# docs/swagger.yaml
basePath: /
definitions:
  models.URL:
    properties:
      alias:
        type: string
      created_at:
        type: string
      id:
        type: integer
      url:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  urlhandler.Request:
    properties:
      alias:
        type: string
      url:
        type: string
    required:
    - url
    type: object
  urlhandler.Response:
    properties:
      alias:
        type: string
      message:
        type: string
    type: object
  urlhandler.UrlList:
    properties:
      urls:
        description: |-
          TotalCount uint64        `json:"total_count"`
          TotalPages uint64        `json:"total_pages"`
          Page       uint64        `json:"page"`
          Size       uint64        `json:"size"`
        items:
          $ref: '#/definitions/models.URL'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server celler server.
  title: Tiny URL API
  version: "1.0"
paths:
  /{alias}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Alias of the URL
        in: path
        name: alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Found
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Redirect URL
      tags:
      - URL
  /url:
    post:
      consumes:
      - application/json
      parameters:
      - description: URL
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/urlhandler.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/urlhandler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Save URL
      tags:
      - URL
  /url/{alias}:
    delete:
      consumes:
      - application/json
      description: delete url by alias
      parameters:
      - description: Alias of the URL
        in: path
        name: alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.URL'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete URL
      tags:
      - URL
    get:
      consumes:
      - application/json
      description: get url string by alias
      parameters:
      - description: Alias of the URL
        in: path
        name: alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.URL'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get URL
      tags:
      - URL
  /url/all:
    get:
      consumes:
      - application/json
      description: get urls
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/urlhandler.UrlList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get All URLs
      tags:
      - URL
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"